+++ packages/tagflow_table/lib/tagflow_table.dart
@@ -1,3 +1,4 @@
 export 'src/rendering/tagflow_table.dart';
 export 'src/widgets/converter.dart';
+export 'src/widgets/semantic_registry.dart';
 export 'src/widgets/tagflow_table.dart';